Create README.md
Exercises to make an object class in python. In these exercises, I made a new class, added definitions to it, and tested it. 

The exercise with the car was to create a Car class with two instance attributes:

.color, which stores the name of the car’s color as a string
.mileage, which stores the number of miles on the car as an integer
Then instantiate two Car objects—a blue car with 20,000 miles and a red car with 30,000 miles—and print out their colors and mileage. Your output should look like this:

The blue car has 20,000 miles.
The red car has 30,000 miles.